Rename task script
