+++ components/job-orchestration/job_orchestration/executor/compress/celeryconfig.py
@@ -6,13 +6,13 @@ from job_orchestration.scheduler.constants import QueueName, TASK_QUEUE_HIGHEST_
 # Force workers to consume only one task at a time
 worker_prefetch_multiplier = 1
 imports = [
-    "job_orchestration.executor.compress.fs_compression_task",
+    "job_orchestration.executor.compress.compression_task",
 ]
 
 # Queue settings
 task_queue_max_priority = TASK_QUEUE_HIGHEST_PRIORITY
 task_routes = {
-    "job_orchestration.executor.compress.fs_compression_task.compress": QueueName.COMPRESSION,
+    "job_orchestration.executor.compress.compression_task.compress": QueueName.COMPRESSION,
 }
 task_create_missing_queues = True
